+++ src/index.ts
@@ -12,4 +12,4 @@ governing permissions and limitations under the License.
 
 export * from './contract';
 export * from './types/experience/Experience';
-export { ExperienceManager, ExperienceError } from './types/experience/ExperienceManager';
+export { ExperienceService, ExperienceError } from './types/experience/ExperienceService';
